creating utils.go and separating responsabilities
