+++ pkg/catalogv2/git/git.go
@@ -423,58 +423,3 @@ func (g *git) injectAgent(cmd *exec.Cmd) (io.Closer, error) {
 
 	return l, nil
 }
-
-func formatGitURL(endpoint, branch string) string {
-	u, err := url.Parse(endpoint)
-	if err != nil {
-		return ""
-	}
-
-	pathParts := strings.Split(u.Path, "/")
-	switch u.Hostname() {
-	case "github.com":
-		if len(pathParts) >= 3 {
-			org := pathParts[1]
-			repo := strings.TrimSuffix(pathParts[2], ".git")
-			return fmt.Sprintf("https://api.github.com/repos/%s/%s/commits/%s", org, repo, branch)
-		}
-	case "git.rancher.io":
-		repo := strings.TrimSuffix(pathParts[1], ".git")
-		u.Path = fmt.Sprintf("/repos/%s/commits/%s", repo, branch)
-		return u.String()
-	}
-
-	return ""
-}
-
-func firstField(lines []string, errText string) (string, error) {
-	if len(lines) == 0 {
-		return "", errors.New(errText)
-	}
-
-	fields := strings.Fields(lines[0])
-	if len(fields) == 0 {
-		return "", errors.New(errText)
-	}
-
-	if len(fields[0]) == 0 {
-		return "", errors.New(errText)
-	}
-
-	return fields[0], nil
-}
-
-func formatRefForBranch(branch string) string {
-	return fmt.Sprintf("refs/heads/%s", branch)
-}
-
-type basicRoundTripper struct {
-	username string
-	password string
-	next     http.RoundTripper
-}
-
-func (b *basicRoundTripper) RoundTrip(request *http.Request) (*http.Response, error) {
-	request.SetBasicAuth(b.username, b.password)
-	return b.next.RoundTrip(request)
-}
